+++ src/main/java/com/spotify/docker/client/messages/swarm/EndpointVirtualIp.java
@@ -28,6 +28,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.auto.value.AutoValue;
 
+import javax.annotation.Nullable;
 
 @AutoValue
 @JsonAutoDetect(fieldVisibility = ANY, getterVisibility = NONE, setterVisibility = NONE)
@@ -36,6 +37,7 @@ public abstract class EndpointVirtualIp {
   @JsonProperty("NetworkID")
   public abstract String networkId();
 
+  @Nullable
   @JsonProperty("Addr")
   public abstract String addr();